Use 32-bit file descriptors

+++ newlib/libc/include/sys/reent.h
@@ -155,7 +155,7 @@ struct __sFILE {
   int	_r;		/* read space left for getc() */
   int	_w;		/* write space left for putc() */
   short	_flags;		/* flags, below; this FILE is free if 0 */
-  short	_file;		/* fileno, if Unix descriptor, else -1 */
+  int	_file;		/* fileno, if Unix descriptor, else -1 */
   struct __sbuf _bf;	/* the buffer (at least 1 byte, if !NULL) */
   int	_lbfsize;	/* 0 or -_bf._size, for inline putc */
